rename internal module
